@@ -26,7 +26,7 @@
     "@swissquote/crafty-runner-gulp": "1.18.4",
     "@swissquote/crafty-runner-rollup": "1.18.4",
     "@swissquote/crafty-runner-webpack": "1.18.4",
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react": "18.0.0",
+    "react-dom": "18.0.0"
   }
 }
