@@ -26,9 +26,9 @@
     "ava": "4.2.0",
     "c8": "7.11.0",
     "execa": "5.1.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-test-renderer": "17.0.2",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
+    "react-test-renderer": "18.0.0",
     "rmfr": "2.0.0"
   },
   "devDependencies": {
